Add more tests (#33)
* 1) Moved test_model to test_model.sft. 2) test_model now consists of actual unit tests of the model and components, some of which are currently known to fail. 3) Fixed notebook 2 to work with current code.

* Make new tests test for more things. Plenty of failures rightn now but that's the point.

* Finshed adding battery of new component smoke tests. Most fail right now.

* Fixed a couple of small issues that came up while running new tests.

* Added code to provide a lot more info to use in case of test failure (last 2 tests) to assist with diagnosis.

* Fixes per review

+++ src/laser_generic/births.py
@@ -163,7 +163,7 @@ class Births:
             timing.append(int(delta.total_seconds() * 1_000_000))
         self._metrics.append(timing)
 
-        model.patches.populations[tick + 1, :] += todays_births
+        model.patches.populations[tick + 1, :] += np.uint32(todays_births)
 
         return
